Move parser to new file
